adding sweetalert

+++ src/components/Product.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 // Redux
 import { useDispatch } from 'react-redux';
@@ -14,9 +15,25 @@ const Product = ({ product }) => {
     const confirmDeleteProduct = id => {
 
         // ask the user
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'Cancel'
+          }).then((result) => {
+            if (result.value) {
+                
+                // pass it to action
+                dispatch( deleteProductAction(id) );
 
-        // pass it to action
-        dispatch( deleteProductAction(id) );
+            }
+          })
+
+
 
     }
